Internal

Public APIs

TestParticle.prepare — Method

prepare(E, B, F; species=Proton, q=1.0, m=1.0)

Return a tuple consists of particle charge, mass for a prescribed species, analytic EM field functions, and external force F.

TestParticle.prepare — Method

prepare(E, B; species=Proton, q=1.0, m=1.0)

Return a tuple consists of particle charge, mass for a prescribed species and analytic EM field functions. Prescribed species are Electron and Proton; other species can be manually specified with species=Ion/User, q and m.

TestParticle.prepare — Method

prepare(grid, E, B, F; species=Proton, q=1.0, m=1.0)

Return a tuple consists of particle charge, mass for a prescribed species, analytic EM field functions, and external force F.

TestParticle.prepare — Method

prepare(grid, E, B; species=Proton)

Return a tuple consists of particle charge, mass for a prescribed species and interpolated EM field functions.

TestParticle.trace_analytic! — Method

ODE equations for charged particle moving in static analytical EM field.

TestParticle.trace_analytic_relativistic! — Method

ODE equations for relativistic charged particle moving in static analytical EM field.

TestParticle.trace_numeric! — Method

ODE equations for charged particle moving in static numerical EM field.

Private types and methods

TestParticle.sph2cart — Method

Convert from spherical to Cartesian coordinates vector.

TestParticle.trace_analytic_full! — Method

ODE equations for charged particle moving in static analytical EM field and external force field.

TestParticle.trace_numeric_full! — Method

ODE equations for charged particle moving in static numerical EM field and external force field.

TestParticle.trace_numeric_relativistic! — Method

ODE equations for relativistic charged particle moving in static numerical EM field.