Internal

Public APIs

TestParticle.prepare — Method

prepare(grid::CartesianGrid, E, B; species=Proton) -> (q, m, E, B)

Return a tuple consists of particle charge, mass for a prescribed species and interpolated EM field functions.

prepare(grid::CartesianGrid, E, B, B₀::Real; species=Proton) -> (Ω, E, B)

Return a tuple consists of gyrofrequency for normalization and interpolated EM field functions given magnetic field scale B₀ in Tesla.

prepare(grid::CartesianGrid, E, B, F; species=Proton, q=1.0, m=1.0) -> (q, m, E, B, F)

Return a tuple consists of particle charge, mass for a prescribed species of charge q and mass m, interpolated EM field functions, and external force F.

prepare(x::AbstractRange, y::AbstractRange, z::AbstractRange, E, B) -> (q, m, E, B)

Direct range input for uniform grid in 3D is also accepted.

prepare(E, B; species=Proton, q=1.0, m=1.0) -> (q, m, E, B)

Return a tuple consists of particle charge, mass for a prescribed species of charge q and mass m and analytic EM field functions. Prescribed species are Electron and Proton; other species can be manually specified with species=Ion/User, q and m.

prepare(E, B, F; species=Proton, q=1.0, m=1.0) -> (q, m, E, B, F)

Return a tuple consists of particle charge, mass for a prescribed species of charge q and mass m, analytic EM field functions, and external force F.

TestParticle.trace! — Method

trace!(dy, y, p::TPTuple, t)
trace!(dy, y, p::FullTPTuple, t)

ODE equations for charged particle moving in static EM field with in-place form.

ODE equations for charged particle moving in static EM field and external force field with in-place form.

TestParticle.trace — Method

trace(y, p::TPTuple, t) -> SVector{6, Float64}
trace(y, p::FullTPTuple, t) -> SVector{6, Float64}

ODE equations for charged particle moving in static EM field with out-of-place form.

ODE equations for charged particle moving in static EM field and external force field with out-of-place form.

TestParticle.trace_normalized! — Method

trace_normalized!(dy, y, p::TPNormalizedTuple, t)

Normalized ODE equations for charged particle moving in static EM field with in-place form. If the field is in 2D X-Y plane, periodic boundary should be applied for the field in z via the extrapolation function provided by Interpolations.jl.

TestParticle.trace_relativistic! — Method

trace_relativistic!(dy, y, p::TPTuple, t)

ODE equations for relativistic charged particle moving in static EM field with in-place form.

TestParticle.trace_relativistic — Method

trace_relativistic(y, p::TPTuple, t) -> SVector{6, Float64}

ODE equations for relativistic charged particle moving in static EM field with out-of-place form.

TestParticleMakie.monitor — Method

monitor(sol::AbstractODESolution; vars=nothing, tspan=nothing)

A plot recipe for monitor the orbit of a particle and other physics quantities.

Arguments

sol::AbstractODESolution: the solution returned by the ODE solver.

Keywords

vars: the argument used to choose the variables to be plotted. Default value is [4, 5, 6].

tspan::Tuple: the span of time to be plotted. For example, tspan = (0, 1).

vars

The usage of vars for this function is different from those in orbit. It can only be a list and its length is equal to 3. The type of its elements must be Integer or Function. The form of a function is the same as those prescribed by orbit. For example,

Eₖ(xu) = mₑ*(xu[4]^2 + xu[5]^2 + xu[6]^2)/2
monitor(sol, vars=[1, 2, Eₖ])

TestParticleMakie.orbit — Method

orbit(sol::AbstractODESolution; vars=nothing, tspan=nothing, to_3d::Bool=false, interactive::Bool=true)

A plot recipe for plotting orbit or other figures related to six phase space coordinates and time.

Arguments

sol::AbstractODESolution: the solution returned by the ODE solver.

Keywords

vars: the argument used to choose variables to be plotted. Default value is [(1, 2, 3)].

tspan::Tuple: the span of time to be plotted. For example, tspan = (0, 1).

to_3d::Bool: whether to force the points to be plotted in 3d. If true, the order of coordinates will be rearranged for plotting as 3d points.

interactive::Bool: whether to show the figure in interactive mode.

vars

vars can be an Integer, Function, Tuple and Array. The basic form for vars is Tuple. For example, if you want to plot the variable x as a function of time or the orbit of a particle, you can use it like this:

orbit(sol, vars=(0, 1))
orbit(sol, vars=(1, 2, 3))

If the length of the tuple is 3, it will be converted to 3d. If you want to plot a function of time, position or velocity, you can first define a function. The arguments of the function must be a 7-dimensional array, the elements of which correspond to the phase space coordinates and time, and the return value must be a Number. For example,

Eₖ(xu) = mₑ*(xu[4]^2 + xu[5]^2 + xu[6]^2)/2
orbit(sol, vars=(0, Eₖ))

This command will plot the kinetic energy as a function of x. The above form is equivalent to

orbit(sol, vars=Eₖ)
orbit(sol, vars=1)

Plotting multiple lines at one time are supported by providing a tuple of indexes. For example,

orbit(sol, vars=[1, (1, 2), Eₖ])

There are some more advanced usages. If a tuple contains vectors, they will be expanded automatically. For example,

orbit(sol, vars=(0, [1, 2, 3]))
orbit(sol, vars=([1, 2, 3], [4, 5, 6]))

are equivalent to

orbit(sol, vars=[(0, 1), (0, 2), (0, 3)])
orbit(sol, vars=[(1, 4), (2, 5), (3, 6)])

Private types and methods

TestParticle.Field — Type

Field{itd, F} <: AbstractField{itd}

A representation of a field function f, defined by:

time-independent field

\[\mathbf{F} = F(\mathbf{x}),\]

time-dependent field

\[\mathbf{F} = F(\mathbf{x}, t).\]

Arguments

field_function::Function: the function of field.

itd::Bool: whether the field function is time dependent.

F: the type of field_function.

TestParticle.FullTPTuple — Type

The type of parameter tuple for full test particle problem.

TestParticle.Species — Type

Type for the particles, Proton, Electron, Ion, or User.

TestParticle.TPNormalizedTuple — Type

The type of parameter tuple for normalized test particle problem.

TestParticle.TPTuple — Type

The type of parameter tuple for normal test particle problem.

TestParticle.dipole — Method

Calculates the magnetic field from a dipole with magnetic moment M at r.

TestParticle.fieldline — Function

fieldline(L, ϕ, nP)

Creates points on one field line of the magnetic field from a dipole. In a centered dipole magnetic field model, the path along a given L shell can be described as r = L*cos²λ, where r is the radial distance (in planetary radii) to a point on the line, λ is its co-latitude, and L is the L-shell of interest.

TestParticle.getB_CS_harris — Method

getB_CS_harris(B₀, L)

Return the magnetic field at location r near a current sheet with magnetic strength B₀ and sheet length L.

TestParticle.getB_bottle — Method

getB_bottle(x, y, z, distance, a, b, I1, I2) -> Vector{Float}

Get magnetic field from a magnetic bottle. Reference: https://en.wikipedia.org/wiki/Magneticmirror#Magneticbottles

Arguments

x,y,z::Float: location in [m].

distance::Float: distance between solenoids in [m].

a::Float: radius of each side coil in [m].

b::Float: radius of central coil in [m].

I1::Float: current in the solenoid times number of windings in side coils.

I2::Float: current in the central solenoid times number of windings in the

central loop.

TestParticle.getB_dipole — Method

Analytic magnetic field function for testing. Return in SI unit.

TestParticle.getB_tokamak_coil — Method

getB_tokamak_coil(x, y, z, a, b, ICoils, IPlasma)

Get the magnetic field from a Tokamak topology consists of 16 coils. Original: https://github.com/BoschSamuel/Simulation-of-a-Tokamak-Fusion-Reactor/blob/master/Simulation2.m

Arguments

x,y,z::Float: location in [m].

a::Float: radius of each coil in [m].

b::Float: radius of central region in [m].

ICoil::Float: current in the coil times number of windings.

IPlasma::Float: current of the plasma?

TestParticle.getB_tokamak_profile — Method

getB_tokamak_profile(x, y, z, q_profile, a, R₀, Bζ0)

Reconstruct the magnetic field distribution from a safe factor(q) profile. The formulations are from the book "Tokamak 4th Edition" by John Wesson.

Arguments

x,y,z::Float: location in [m].

q_profile::Function: profile of q. The variable of this function must be the normalized radius.

a::Float: minor radius [m].

R₀::Float: major radius [m].

Bζ0::Float: toroidal magnetic field on axis [T].

TestParticle.getE_dipole — Method

Analytic electric field function for testing.

TestParticle.get_gc — Method

get_gc(param::Union{TPTuple, FullTPTuple})

Get three function for plotting the orbit of guiding center.

For example:

param = prepare(E, B; species=Proton)
gc = get_gc(params)
# The definitions of stateinit, tspan, E and B are ignored.
prob = ODEProblem(trace!, stateinit, tspan, param)
sol = solve(prob, Vern7(); dt=2e-11)

orbit(sol, vars=gc)

TestParticle.guiding_center — Method

guiding_center(xu, param::Union{TPTuple, FullTPTuple})

Calculate the coordinates of the guiding center according to the phase space coordinates of a particle. Reference: https://en.wikipedia.org/wiki/Guiding_center

A simple definition:

\[\mathbf{X}=\mathbf{x}-m\frac{\mathbf{v}\times\mathbf{B}}{qB}\]

TestParticle.set_axes_equal — Method

set_axes_equal(ax)

Set 3D plot axes to equal scale for Matplotlib. Make axes of 3D plot have equal scale so that spheres appear as spheres and cubes as cubes. Required since ax.axis('equal') and ax.set_aspect('equal') don't work on 3D.

TestParticle.sph2cart — Method

Convert from spherical to Cartesian coordinates vector.