Internal

Public APIs

TestParticle.prepare — Method

prepare(grid, E, B, species="proton")

Return a tuple consists of particle charge, mass for a prescribed species and interpolated EM field functions.

TestParticle.prepare — Method

prepare(E, B; species="proton", q=1.0, m=1.0)

Return a tuple consists of particle charge, mass for a prescribed species and analytic EM field functions. Prescribed species are "electron" and "proton"; other species can be manually specified with species="other", q and m.

TestParticle.trace_analytic! — Method

ODE equations for charged particle moving in static analytical EM field.

TestParticle.trace_analytic_relativistic! — Method

ODE equations for relativistic charged particle moving in static analytical EM field.

TestParticle.trace_numeric! — Method

ODE equations for charged particle moving in static numerical EM field.

Private types and methods

TestParticle.sph2cart — Method

Convert from spherical to Cartesian coordinates vector.